\name{getPathThroughRegions}
\alias{getPathThroughRegions}
\title{Naive traveling salesman heuristic without return}
\description{Finds a minimum length Hamiltonian path through the sampled regions of a tumor, given their coordinates.}
\usage{getPathThroughRegions(coord)}
\arguments{
  \item{coord}{
     Matrix with one row per sample and two columns containing the sample's x- and y-coordinates.
  }
}
\details{
}

\author{Noemi Andor}


\examples{}




