\name{getPedigreeTree}
\alias{getPedigreeTree}
\title{Phylogeny depicting how cell line lineages are related.}
\description{Visualizes relations between lineages of a cell line.}
\usage{getPedigreeTree(cellLine = cellLine, id = NULL)}
\arguments{
  \item{cellLine}{
     Name of the cell line for which to plot the lineage tree. If set to NULL, next argument will be used.
  }
  \item{id}{
    ID of the lineage for which to plot the tree of all derived sublineages. This must be an existing key in the Passaging table.
  }
}
\details{
	
}

\author{Noemi Andor}


