\name{getState}
\alias{getState}
\title{Retrieving state associated with a clone}
\description{The state in which a clone was sequenced}
\usage{getState(cloneID, whichP="TranscriptomePerspective"}
\arguments{
  \item{cloneID}{
     Clone ID (integer).
  }
  \item{whichP}{
     What to vizualize: GenomePerspective (default), TranscriptomePerspective or Identity.
  }
 
}
\details{
}
\value{A string describing subclones' state.
}


